plant_watering/requirements.txt
ipython
clipspy
